Moved constants to their own file
