+++ services/constants.js
@@ -1,0 +1,30 @@
+const Festival = {
+    LEGEND : "legend",
+    FLASH  : "flash"
+}
+
+const Season = {
+    VALENTINE : "valentine",
+    SUMMER    : "summer",
+    HALLOWEEN : "halloween",
+    HOLIDAY   : "holiday"
+}
+
+const ItemType = {
+	WEAPON : 0,
+	SUMMON : 1
+}
+
+const Rarity = {
+	R   : 1,
+	SR  : 2,
+	SSR : 3
+}
+
+const SSRRate = 3.0
+
+exports.ItemType = ItemType
+exports.Festival = Festival
+exports.Rarity = Rarity
+exports.Season = Season
+exports.SSRRate = SSRRate
